django/privat_gos_sait/requirements.txt
amqp==2.5.0
billiard==3.6.0.0
bleach==3.1.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-grappelli==2.13.1
django-markdown==0.8.4
django-markdown-filter==0.0.1
django-markdownx==2.0.28
django-paypal==1.0.0
django-summernote==0.8.11.4
idna==2.8
kombu==4.6.3
Markdown==3.1.1
markdown2==2.3.8
oauthlib==3.0.1
paypal==1.2.5
Pillow==6.0.0
psycopg2==2.8.3
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
webencodings==0.5.1
